fix(module: tree): fix wrong checked nodes
